requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
et_xmlfile==2.0.0
fastapi==0.119.0
fastapi-utils==0.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
psutil==5.9.8
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.37.0
